platform: x86: Stylistic updates for intel-skl-int3472
This commit makes a bunch of stylistic updates, minor changes and other
stuff that's part of the improvements pass I'm doing to the code after
taking into account feedback from the list.

It also alters the ACPI buffer fetching code to be more generalisable so
I can re-use it to fetch the clock frequency.

Signed-off-by: Daniel Scally <djrscally@gmail.com>
Patchset: cameras