media: i2c: update set_fmt() for ov5693
The set_fmt() function is a bit messy still, using home grown solutions to
find the closest supported resolution instead of the v4l2 helpers. It also
fails to update control ranges to account for the new mode (though this is
moot currently as they're all the same, but the probably shouldn't be).

Fix it up a little.

Signed-off-by: Daniel Scally <djrscally@gmail.com>
Patchset: cameras