Add info for the EZpad 6s Pro
An user came to BlissOS Telegram chat asked to add this in. The
tablet can be able to use values from the Pro B model. Different
from other Ezpad 6 Pro, this one has a BIOS that was made in 2020
and "5.12" is the BIOS revision, it has a different string for
the BIOS version.

Signed-off-by: hmtheboy154 <buingoc67@gmail.com>