ANDROID: sched: Avoid re-evaluating proxy_needs_return when sched_proxy_exec() is disabled
With proxy-execution, we add an extra check such that should the task
have already been woken up, but it is still in the BO_WAKING state,
thus potentially needed return-migration, we avoid breaking out of
try_to_wakeup() in order so that the proxy_needs_return() logic in
ttwu_runnable() can dequeue the task and allow it to be woken on its
wake_cpu (effecively doing the return migration).

As reported kuyo chang, this check is not conditionalized by
sched_proxy_exec(), so when proxy-execution is disabled its possible
for us to hit this case, and continue to ttwu_runnable(), however
we won't do the proxy_needs_return() logic as it is conditionalized,
resulting in the enqueued task being enqueued again, triggering
warnings.

WARNING: CPU: 5 PID: 20337 at kernel/sched/rt.c:1136 dequeue_rt_stack+0x1a0/0x26c
Call trace:
 dequeue_rt_stack+0x1a0/0x26c
 enqueue_task_rt+0xa0/0x39c
 enqueue_task+0x7c/0x340
 do_activate_task+0xcc/0x178
 ttwu_do_activate+0xac/0x2a8
 try_to_wake_up+0x748/0xb84
 wake_up_q+0x70/0xd8
...

or

WARNING: CPU: 6 PID: 0 at kernel/sched/rt.c:1823 pick_task_rt+0x98/0xb0
Call trace:
 pick_task_rt+0x98/0xb0
 __schedule+0x494/0xc8c
 schedule_idle+0x24/0x48
 do_idle+0x248/0x290
...

Thus, pull the check out to a separate function and include the
check on sched_proxy_exec() so that it is properly conditionalized.

Bug: 388983726
Fixes: e4239ea599c6 ("ANDROID: sched: Handle blocked-waiter migration (and return migration)")
Change-Id: Id21c810b240d54aafc63bd0c34255eabdbefe94a
Reported-by: kuyo chang <kuyo.chang@mediatek.com>
Suggested-by: kuyo chang <kuyo.chang@mediatek.com>
Signed-off-by: John Stultz <jstultz@google.com>